add vuestic ui
