+++ candle/package.json
@@ -11,7 +11,8 @@
     "cometd": "^7.0.5",
     "cometd-nodejs-client": "^1.3.0",
     "pinia": "^2.0.9",
-    "vue": "^3.2.25"
+    "vue": "^3.2.25",
+    "vuestic-ui": "^1.3.3"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^2.0.0",
