+++ candle/src/App.vue
@@ -1,12 +1,11 @@
 <script setup lang="ts">
 // This starter template is using Vue 3 <script setup> SFCs
 // Check out https://v3.vuejs.org/api/sfc-script-setup.html#sfc-script-setup
-import HelloWorld from './components/HelloWorld.vue'
-import HelloAntDesign from "./components/HelloAntDesign.vue";
+import HelloVuesticUi from "./components/HelloVuesticUi.vue";
 </script>
 
 <template>
-  <HelloAntDesign />
+  <HelloVuesticUi />
 </template>
 
 <style>
